Remove 'sunbeam inspect' command
Also move 'plans' and 'unlock-plan' to utils.
